requirements/base.txt
Django==2.0.4
djangorestframework==3.8.2
djangorestframework-xml
django-cors-headers
django-redis
wheel
pycrypto
mysqlclient
requests
python-dateutil
pytest
python-alipay-sdk
wechatpy
numpy
matplotlib
gunicorn
incremental>=16.10.1
redis==2.10.6
channels==1.0.2
daphne==1.3.0
asgiref~=1.1
asgi_redis==1.4.3
pypng
pyqrcode